requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
Pillow==10.4.0
celery==5.3.4
redis==5.0.1
firebase-admin==6.2.0
django-extensions==3.2.3
python-decouple==3.8
